+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/application/AbstractBinderFactorySpec.groovy
@@ -433,7 +433,7 @@ public class AbstractBinderFactorySpec extends Specification {
 
         when:
         systemConfig.setProperty(DIMENSION_BACKEND_KEY, "memory")
-        ConfigurationLoader cl = localBinderFactory.buildConfigurationLoader()
+        ConfigurationLoader cl = localBinderFactory.buildConfigurationLoader(Mock(DataSourceMetadataService))
 
         then:
         cl.dimensionLoader instanceof TypeAwareDimensionLoader