+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/data/DruidQueryBuilderSpec.groovy
@@ -28,6 +28,7 @@ import com.yahoo.bard.webservice.druid.model.query.DruidAggregationQuery
 import com.yahoo.bard.webservice.druid.model.query.GroupByQuery
 import com.yahoo.bard.webservice.druid.model.query.TimeSeriesQuery
 import com.yahoo.bard.webservice.druid.model.query.TopNQuery
+import com.yahoo.bard.webservice.metadata.DataSourceMetadataService
 import com.yahoo.bard.webservice.table.ConcretePhysicalTable
 import com.yahoo.bard.webservice.table.PhysicalTable
 import com.yahoo.bard.webservice.table.resolver.DefaultPhysicalTableResolver
@@ -130,7 +131,13 @@ class DruidQueryBuilderSpec extends Specification {
     def "Test recursive buildQueryMethods"() {
         setup:
         Set apiSet = (["abie1234", "abde1129"].collect() { apiFilters.get(it) }) as Set
-        PhysicalTable tab = new ConcretePhysicalTable("tab1", DAY.buildZonedTimeGrain(UTC), [] as Set, [:])
+        PhysicalTable tab = new ConcretePhysicalTable(
+                "tab1",
+                DAY.buildZonedTimeGrain(UTC),
+                [] as Set,
+                [:],
+                Mock(DataSourceMetadataService)
+        )
         Filter filter = FILTER_BUILDER.buildFilters([(resources.d3): apiSet])
         ZonedTimeGrain granularity = WEEK.buildZonedTimeGrain(UTC)
         Set dimension = [resources.d1] as Set
@@ -229,7 +236,8 @@ class DruidQueryBuilderSpec extends Specification {
                 "tab1",
                 DAY.buildZonedTimeGrain(UTC),
                 [] as Set,
-                [:]
+                [:],
+                Mock(DataSourceMetadataService)
         )
         Filter filter = FILTER_BUILDER.buildFilters([(resources.d3): apiSet])
         ZonedTimeGrain granularity = YEAR.buildZonedTimeGrain(UTC)